+++ strawberryfields/backends/tfbackend/ops.py
@@ -1077,11 +1077,14 @@ def partial_trace(system, mode, state_is_pure, batched=False):
     return reduced_state
 
 
-def reduced_density_matrix(system, mode, state_is_pure, batched=False):
+def reduced_density_matrix(system, modes, state_is_pure, batched=False):
     """
-    Trace out all subsystems except 'mode' from 'system'.
+    Trace out all subsystems except those specified in ``modes`` from ``system``. ``modes`` can be either an int or a list.
     This operation always returns a mixed state, since we do not know in advance if a mode is entangled with others.
     """
+    if isinstance(modes, int):
+        modes = [modes]
+
     if state_is_pure:
         reduced_state = mixed(system, batched)
     else:
@@ -1094,7 +1097,7 @@ def reduced_density_matrix(system, mode, state_is_pure, batched=False):
     num_modes = (num_indices - batch_offset) // 2  # always mixed
     removed_cnt = 0
     for m in range(num_modes):
-        if m != mode:
+        if m not in modes:
             reduced_state = partial_trace(reduced_state, m - removed_cnt, False, batched)
             removed_cnt += 1
     return reduced_state